+++ immersed-presets-decklink/src/testFixtures/java/org/bytedeco/decklink/windows/IUnknownSupport.java
@@ -28,7 +28,7 @@ public interface IUnknownSupport
 
     long AddRef();
 
-    int QueryInterface(GUID riid, PointerPointer ppv);
+    long QueryInterface(GUID riid, PointerPointer ppv);
 
     long Release();
 }
