+++ immersed-presets-decklink/src/testFixtures/java/org/bytedeco/decklink/windows/capturepreview/DeckLinkDiscovery.java
@@ -81,7 +81,7 @@ public class DeckLinkDiscovery extends IDeckLinkDeviceNotificationCallback
     }
 
     @Override
-    public int QueryInterface(GUID riid, PointerPointer ppv)
+    public long QueryInterface(GUID riid, PointerPointer ppv)
     {
         return m_com.QueryInterface(riid, ppv);
     }
